requirements.txt
langchain
langchain-openai
pydantic<2
sqlalchemy
psycopg2-binary
python-dotenv
pytest 